requirements.txt
amulet-core
numpy